requirements-dev.txt
# Pytest
pytest-flask==1.2.0

# CodeStyle
pycodestyle==2.7.0

# Flake8
flake8==3.9.2

# Coverage
coverage==5.5

-r requirements.txt
